reorganize client rpc structure
