+++ src/client/rpc/mod.rs
@@ -1,0 +1,123 @@
+use crate::errors::NodeRpcClientError;
+use async_trait::async_trait;
+use core::fmt;
+use crypto::merkle::{MerklePath, MmrDelta};
+use objects::{
+    accounts::AccountId,
+    notes::{NoteId, NoteMetadata},
+    transaction::ProvenTransaction,
+    BlockHeader, Digest,
+};
+
+mod tonic_client;
+pub use tonic_client::TonicRpcClient;
+
+// NODE API TRAIT
+// ================================================================================================
+
+#[async_trait]
+pub trait NodeRpcClient {
+    fn new(config_endpoint: &str) -> Self;
+    async fn submit_proven_transaction(
+        &mut self,
+        proven_transaction: ProvenTransaction,
+    ) -> Result<(), NodeRpcClientError>;
+    async fn get_block_header_by_number(
+        &mut self,
+        block_number: Option<u32>,
+    ) -> Result<BlockHeader, NodeRpcClientError>;
+    async fn sync_state(
+        &mut self,
+        block_num: u32,
+        account_ids: &[AccountId],
+        note_tags: &[u16],
+        nullifiers_tags: &[u16],
+    ) -> Result<StateSyncInfo, NodeRpcClientError>;
+}
+
+// STATE SYNC INFO
+// ================================================================================================
+
+/// Represents a [SyncStateResponse] with fields converted into domain types
+pub struct StateSyncInfo {
+    /// The block number of the chain tip at the moment of the response
+    pub chain_tip: u32,
+    /// The returned block header
+    pub block_header: BlockHeader,
+    /// MMR delta that contains data for (current_block.num, incoming_block_header.num-1)
+    pub mmr_delta: MmrDelta,
+    /// Tuples of AccountId alongside their new account hashes
+    pub account_hash_updates: Vec<(AccountId, Digest)>,
+    /// List of tuples of Note ID, Note Index and Merkle Path for all new notes
+    pub note_inclusions: Vec<CommittedNote>,
+    /// List of nullifiers that identify spent notes
+    pub nullifiers: Vec<Digest>,
+}
+
+// COMMITTED NOTE
+// ================================================================================================
+
+/// Represents a committed note, returned as part of a [SyncStateResponse]
+pub struct CommittedNote {
+    /// Note ID of the committed note
+    note_id: NoteId,
+    /// Note index for the note merkle tree
+    note_index: u32,
+    /// Merkle path for the note merkle tree up to the block's note root
+    merkle_path: MerklePath,
+    /// Note metadata
+    metadata: NoteMetadata,
+}
+
+impl CommittedNote {
+    pub fn new(
+        note_id: NoteId,
+        note_index: u32,
+        merkle_path: MerklePath,
+        metadata: NoteMetadata,
+    ) -> Self {
+        Self {
+            note_id,
+            note_index,
+            merkle_path,
+            metadata,
+        }
+    }
+
+    pub fn note_id(&self) -> &NoteId {
+        &self.note_id
+    }
+
+    pub fn note_index(&self) -> u32 {
+        self.note_index
+    }
+
+    pub fn merkle_path(&self) -> &MerklePath {
+        &self.merkle_path
+    }
+
+    #[allow(dead_code)]
+    pub fn metadata(&self) -> NoteMetadata {
+        self.metadata
+    }
+}
+
+// RPC API ENDPOINT
+// ================================================================================================
+//
+#[derive(Debug)]
+pub enum RpcApiEndpoint {
+    GetBlockHeaderByNumber,
+    SyncState,
+    SubmitProvenTx,
+}
+
+impl fmt::Display for RpcApiEndpoint {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            RpcApiEndpoint::GetBlockHeaderByNumber => write!(f, "get_block_header_by_number"),
+            RpcApiEndpoint::SyncState => write!(f, "sync_state"),
+            RpcApiEndpoint::SubmitProvenTx => write!(f, "submit_proven_transaction"),
+        }
+    }
+}
