+++ src/store/mod.rs
@@ -51,7 +51,7 @@ pub mod tests {
         mock::{MockDataStore, MockRpcApi},
     };
 
-    use crate::client::NodeRpcClient;
+    use crate::client::rpc::NodeRpcClient;
 
     use super::{migrations, Store};
